% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_component_score.R
\name{calculate_component_score}
\alias{calculate_component_score}
\title{Calculate Component Score}
\usage{
calculate_component_score(indicator_score_table)
}
\arguments{
\item{indicator_score_table}{A data frame representing the score table containing metric scores and suitability.}
}
\value{
A data frame containing calculated component scores and grades.
}
\description{
This internal function calculates component scores based on the provided score table.
}
