% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_grade.R
\name{calculate_grade}
\alias{calculate_grade}
\title{This function calculates the grade based on a numeric input value.}
\usage{
calculate_grade(x)
}
\arguments{
\item{x}{A numeric vector representing a score from 0-1.}
}
\value{
A character vector indicating the corresponding grades.
}
\description{
This function calculates the grade based on a numeric input value.
}
\examples{
calculate_grade(0.85) # Returns "A"
calculate_grade(0.6) # Returns "B+"
calculate_grade(0.3) # Returns "C-"
}
