% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_health_score.R
\name{calculate_health_score}
\alias{calculate_health_score}
\title{Calculate  Health Score}
\usage{
calculate_health_score(
  indicators,
  reference_table,
  npsfm_only = FALSE,
  overall_score = TRUE
)
}
\arguments{
\item{indicators}{A list of data frames containing information about indicators. Each data frame should have  data about a different indicator and the columns: "site", "class", "indicator", "component", and "reporting_scale", plus extra columns for metrics measured.}

\item{reference_table}{A data frame providing default values for excellent and poor ecological conditions. If the reference_table user input is not the one provided with the package (reference_table_default), it will be flagged as user defined benchmarks.}

\item{npsfm_only}{Logical indicating whether to consider only NPSFM components (default is FALSE).}

\item{overall_score}{Logical indicating whether to calculate overall scores or group by reporting scale (default is TRUE).}
}
\description{
This function is a wrapper to harmonize and integrate metric data into indicator and component health scores as outlined by Clapcott et al. 2019
}
\examples{
 macroinvertebrates <- data.frame( site = 1:3,class = c("universal"),
 indicator = "macroinvertebrates",
 component = "aquatic_life",
 reporting_scale = c("region_1","region_2","region_1"),
 mci = c(89,120,135))

 calculate_health_score(indicators = list(macroinvertebrates),
 reference_table = reference_table_default)
}
