% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_calculate_indicator_score.R
\name{calculate_indicator_score}
\alias{calculate_indicator_score}
\title{Calculate Indicator Score}
\usage{
calculate_indicator_score(score_table)
}
\arguments{
\item{score_table}{A data frame representing the score table containing metric scores and suitability.}
}
\value{
A data frame containing calculated indicator scores and grades aggregated at the indicator and reporting scale level
}
\description{
This function calculates indicator scores based on the provided score table.
}
