% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_nonlinear_performance_score.R
\name{calculate_nonlinear_performance_score}
\alias{calculate_nonlinear_performance_score}
\title{Calculate Nonlinear Performance Score}
\usage{
calculate_nonlinear_performance_score(
  value,
  reference,
  bottom_line,
  reference_range,
  bottom_line_range
)
}
\arguments{
\item{value}{A numeric vector representing the actual values.}

\item{reference}{A numeric vector representing the reference values.}

\item{bottom_line}{A numeric vector representing the bottom line values.}

\item{reference_range}{A numeric vector representing the reference range upper bounds.}

\item{bottom_line_range}{A numeric vector representing the bottom line range lower bounds.}
}
\value{
A numeric vector of the same length as 'value', representing the calculated performance scores.
}
\description{
This function calculates a nonlinear performance score based on provided values, reference ranges, and bottom lines.
}
