% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_npsfm_grade.R
\name{calculate_npsfm_grade}
\alias{calculate_npsfm_grade}
\title{Calculate NPSFM Grades}
\usage{
calculate_npsfm_grade(indicators, reference_table, overall_score = TRUE)
}
\arguments{
\item{indicators}{A list of data frames containing information about indicators. Each data frame should have columns: "site", "class", "indicator", "component", and "reporting_scale", plus extra columns for metrics measured.}

\item{reference_table}{A data frame providing default values for excellent and poor ecological conditions. If the reference_table user input is not the one provided with the package (reference_table_default), it will be flagged as user defined benchmarks.}

\item{overall_score}{Logical indicating whether to calculate overall scores or group by reporting scale (default is TRUE).}
}
\description{
This function calculates NPSFM  grades  based on mean metric values and grade bands defined in the 2020 National Policy Statement for Freshwater Management.
}
\examples{
 macroinvertebrates <- data.frame( site = 1:3,class = c("universal"),
 indicator = "macroinvertebrates",
 component = "aquatic_life",
 reporting_scale = c("region_1","region_2","region_1"),
 mci = c(89,120,135))

 calculate_npsfm_grade(indicators = list(macroinvertebrates),
 reference_table = reference_table_default,
 overall = TRUE)
}
