% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_plot_width.R
\name{calculate_plot_width}
\alias{calculate_plot_width}
\title{Calculate availability score for plotting}
\usage{
calculate_plot_width(
  reference_table,
  score_table,
  npsfm_only = TRUE,
  all_components = FALSE
)
}
\arguments{
\item{reference_table}{Reference table containing information on suitability of data collection.}

\item{score_table}{Score table containing actual data collected.}

\item{npsfm_only}{Logical indicating whether to consider only NPSFM components (default is TRUE).}

\item{all_components}{Logical indicating whether to calculate availability for all components
(default is FALSE). If FALSE, availability is calculated for each component
and its indicators. Users should not have to alter this argument.}
}
\value{
A data frame with availability scores calculated for each component and indicator or reporting scale.
}
\description{
This function calculates the availability score based on the suitability of collected data
compared to potential data that could have been collected, for each component and indicator.
}
