% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_suitability_score.R
\name{calculate_suitability_score}
\alias{calculate_suitability_score}
\title{Calculate Suitability Weighted Score}
\usage{
calculate_suitability_score(
  performance_table,
  npsfm_only = FALSE,
  overall_score = TRUE
)
}
\arguments{
\item{performance_table}{A data frame representing the performance table containing nested metric performance scores.}

\item{npsfm_only}{Logical indicating whether to consider only NPSFM components (default is FALSE).}

\item{overall_score}{Logical indicating whether to calculate overall scores or group by reporting scale (default is TRUE).}
}
\description{
This function calculates suitability scores based on the provided performance table and optional suitability table.
}
\examples{
 macroinvertebrates <- data.frame( site = 1:3,class = c("universal"),
 indicator = "macroinvertebrates",
 component = "aquatic_life",
 reporting_scale = c("region_1","region_2","region_1"),
 mci = c(89,120,135))

 component_table <- make_component_table(indicators = list(macroinvertebrates))

 performance_scores <- calculate_performance_scores(component_table = component_table,
  reference_table = reference_table_default)

calculate_suitability_score(performance_table = performance_scores,
npsfm_only = FALSE,overall_score = TRUE)

}
