% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_component_table.R
\name{check_component_table}
\alias{check_component_table}
\title{Check validity of component table}
\usage{
check_component_table(component_table)
}
\arguments{
\item{component_table}{A data frame containing information about a single component, with columns:
'indicator', 'component', and 'data'.}
}
\value{
The input component_table if valid; otherwise, raises an error with appropriate message.
}
\description{
This function checks the validity of a provided component table to ensure it meets necessary
requirements for further processing.
}
