% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_indicator_table.R
\name{check_indicator_table}
\alias{check_indicator_table}
\title{Check Indicator Table}
\usage{
check_indicator_table(indicator_table, key_columns)
}
\arguments{
\item{indicator_table}{The indicator table to be validated (data frame).}

\item{key_columns}{A vector of column names considered key columns.}
}
\value{
The input indicator table if validation passes.
}
\description{
This function validates an indicator table and identifies metrics.
}
