% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_reference_table.R
\name{check_reference_table}
\alias{check_reference_table}
\title{Check Reference Table}
\usage{
check_reference_table(reference_table, key_columns)
}
\arguments{
\item{reference_table}{The reference table to be validated (data frame). Can be the default provided with the package or provided by the user. If provided by the user, it should have the same col um names as the default.}

\item{key_columns}{A vector of column names considered key columns.}
}
\value{
The input reference table if validation passes.
}
\description{
This function validates a reference table.
}
