% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fish}
\alias{fish}
\title{Fish Data to calculate New Zealand River Ecosystem Health Score}
\format{
A data frame with 2824 rows and 6 variables:
\describe{
\item{site}{The unique identifier for the spatial scale at which observations were aggregated.}
\item{class}{The reference category used for calculating metric performance scores.}
\item{indicator}{The specific indicator to which metrics belong (i.e., macroinvertebrates)}
\item{component}{The specific component to which the indicator belongs (i.e.,. aquatic_life)}
\item{reporting_scale}{The chosen scale for data integration}
\item{fish_ibi}{Simulated values of the metric Fish Index of Biological Integrity}
}
}
\usage{
fish
}
\description{
A simulated dataset containing fish observations aggregated at a site level.
}
\keyword{datasets}
