% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{macroinvertebrates}
\alias{macroinvertebrates}
\title{Macroinvertebrates Data to calculate Zealand River Ecosystem Health Score}
\format{
A data frame with 1021 rows and 8 variables:
\describe{
\item{site}{The unique identifier for the spatial scale at which observations were aggregated.}
\item{class}{The reference category used for calculating metric performance scores.}
\item{indicator}{The specific indicator to which metrics belong (i.e., macroinvertebrates)}
\item{component}{The specific component to which the indicator belongs (i.e.,. aquatic_life)}
\item{reporting_scale}{The chosen scale for data integration}
\item{mci}{Simulated values of the metric macroinvertebrate community index (MCI)}
\item{percentage_ept_taxa}{Simulated values of the metric percentage of Ephemeroptera (mayfly), Plecoptera (stonefly) and Trichoptera (caddisfly) taxa. }
\item{ept_taxa_richness}{Simulated values of the richness of EPT taxa.}
}
}
\usage{
macroinvertebrates
}
\description{
A simulated dataset containing macroinvertebrates observations aggregated at a site level.
}
\keyword{datasets}
