% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_component_table.R
\name{make_component_table}
\alias{make_component_table}
\title{Make Component Table}
\usage{
make_component_table(indicators)
}
\arguments{
\item{indicators}{A list of data frames containing information about indicators. Each data frame should have columns: "site", "class", "indicator", "component", and "reporting_scale", plus extra columns for metrics measured.}
}
\value{
A nested data frame where each row is a different indicator and metric data is nested.
}
\description{
This function generates a nested data frame with metrics categorized by indicators.
}
\details{
This function nests metric data within indicators from the same component to aid manipulation of disparate metrics. Data is nested in a long format, where each row is an observation for a given metric, for  a given site. If information is missing from a site, the whole site will be dropped. If information is missing from a metric, only that observation will be dropped.
}
\examples{
# Create example indicator data frames
macroinvertebrates <- data.frame(
site = 1:3,class = c("default", "productive", "default"),
indicator = "macroinvertebrates",
component = "aquatic_life",
reporting_scale = "region",
metric1 = runif(3),
metric2 = runif(3))

fish <- data.frame(
site = 1:3,
class = c("default", "productive", "default"),
indicator = "fish",
component = "aquatic_life",
reporting_scale = "region",
metric1 = runif(3),
metric2 = runif(3))

make_component_table(list(macroinvertebrates, fish))
}
