% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_indicator_table.R
\name{nest_indicator_table}
\alias{nest_indicator_table}
\title{Nest Indicator Table}
\usage{
nest_indicator_table(indicator_table, key_columns)
}
\arguments{
\item{indicator_table}{The indicator table to be processed as a data frame.}

\item{key_columns}{A character vector of column names considered key columns.}
}
\value{
A nested data frame with metrics categorized by components.
}
\description{
This internal function processes an indicator table, pivots it longer, ensures consistent data types and removes NA values for each metric.
}
