% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_reference_table.R
\name{nest_reference_table}
\alias{nest_reference_table}
\title{This function nests a reference table by indicator to facilitate joining with component tables for calculating performance scores.}
\usage{
nest_reference_table(reference_table, key_columns)
}
\arguments{
\item{reference_table}{A data frame representing the reference table to be nested.}

\item{key_columns}{A character vector specifying the key columns necessary for further analyses}
}
\value{
A nested data frame with the reference table grouped by indicator.
}
\description{
This function nests a reference table by indicator to facilitate joining with component tables for calculating performance scores.
}
\details{
This function takes a reference table and nests it based on the indicator and component columns.
It also performs type conversion for character columns to ensure consistency and numeric columns
for compatibility with subsequent calculations.
}
