% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_component.R
\name{plot_component}
\alias{plot_component}
\title{Plot component metrics availability}
\usage{
plot_component(
  score_table,
  reference_table,
  npsfm_only = FALSE,
  color_table = NULL,
  font_size = 20,
  start_year = NULL,
  end_year = NULL
)
}
\arguments{
\item{score_table}{Score table containing metrics for a single component.}

\item{reference_table}{Reference table containing metric information.}

\item{npsfm_only}{Logical indicating whether to consider only NPSFM components (default is FALSE).}

\item{color_table}{Optional color table for customizing plot colors.}

\item{font_size}{Font size for plot labels (default is 20).}

\item{start_year}{Optional start year of data collection (default is NULL).}

\item{end_year}{Optional end year of data collection (default is NULL).}
}
\value{
A list of doughnut plots, each showing the availability and score of indicators for the specified
component and reporting scale.
}
\description{
This function generates doughnut plots showing the availability and grades of indicators for a specific
component based on the provided score table.
}
\examples{
 macroinvertebrates <- data.frame( site = 1:3,class = c("universal"),
 indicator = "macroinvertebrates",
 component = "aquatic_life",
 reporting_scale = c("region_1","region_2","region_1"),
 mci = c(89,120,135))

 score_table <-calculate_health_score(indicators = list(macroinvertebrates),
 reference_table = reference_table_default)

 color_table <- c("Z" = "white",
 "W" = "#bdbcbc",
 "A" = "#20a7ad",
 "B+" = "#85bb5b",
 "B-" = "#85bb5b",
 "C+" = "#ffa827",
  "C-" = "#ffa827",
  "D" = "#e85129")

 plot_component(score_table = score_table,
 reference_table = reference_table_default,
 color_table = color_table,
 start_year = 2017,
 end_year = 2022)
}
