% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_empty.R
\name{plot_empty}
\alias{plot_empty}
\title{Plot empty components for specified reporting scale}
\usage{
plot_empty(
  component_ind,
  reporting_scale,
  reference_table,
  npsfm_only = FALSE,
  color_table = NULL,
  font_size = 20,
  start_year = NULL,
  end_year = NULL
)
}
\arguments{
\item{component_ind}{Component indicator for which metrics availability is to be plotted.}

\item{reporting_scale}{Reporting scale to consider for plotting.}

\item{reference_table}{Reference table containing metric information.}

\item{npsfm_only}{Logical indicating whether to consider only NPSFM components (default is FALSE).}

\item{color_table}{Optional color table for customizing plot colors.}

\item{font_size}{Font size for plot labels (default is 20).}

\item{start_year}{Optional start year of data collection (default is NULL).}

\item{end_year}{Optional end year of data collection (default is NULL).}
}
\value{
A list of doughnut plots, each showing an empty plot for the specified
component and reporting scale.
}
\description{
This function generates doughnut plots showing the no availability of metrics for a specific
component and reporting scale. It visualizes empty plots.
}
