% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_doughnut_chart.R
\name{prepare_doughnut_chart}
\alias{prepare_doughnut_chart}
\title{Prepare Doughnut Chart Data}
\usage{
prepare_doughnut_chart(plot_table)
}
\arguments{
\item{plot_table}{A data frame containing data for plotting doughnut charts.}
}
\value{
A modified data frame with expanded rows suitable for plotting doughnut charts.
}
\description{
This  internal function prepares data for plotting doughnut charts by transforming the input plot_table.
}
