% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_table.R
\docType{data}
\name{reference_table_default}
\alias{reference_table_default}
\title{Reference Table to calculate New Zealand River Ecosystem Health Score}
\format{
A data frame with 60 rows and 14 variables:
\describe{
\item{component}{The component to which a given metric belongs. Each component describes a different aspect of freshwater biophysical ecosystem health}
\item{indicator}{Within a component, indicators aggregate similar metrics. This column denotes the indicator to which a given metric belongs}
\item{attribute}{The biophysical attributes a given metric aims to quantify. This classification often matches the metric classification, but it can vary when multiple metrics measure the same attribute (e.g., both the \emph{do_7_day_min} and \emph{do_1_day_min} metrics measure the attribute \emph{dissolved oxygen})}
\item{metric}{Measured or modeled biophysical characteristics of rivers and streams}
\item{class}{The reference category used for calculating metric performance scores (e.g., productive/default, sediment class)}
\item{bottom_line}{This numerical value is a benchmark that denotes a degraded state for a given metric, against which sites will be compared to calculate performance scores}
\item{bottom_line_range}{For non-linear metrics, these values denote the range that comprises a degraded state for a given metric. If the metric is linear, then this column should be NA}
\item{reference}{This numerical value is a benchmark that denotes a healthy state for a given metric, against which sites will be compared to calculate performance scores}
\item{reference_range}{For non-linear metrics, these values denote the range that comprises a healthy state for a given metric. If the metric is linear, then this column should be NA}
\item{cut}{For NPSFM metrics, this numerical value is the benchmark that denotes metric grades between the \emph{bottom_line} and the \emph{reference} values. These values are only used to calculate NPSFM grade bands and are not used to calculate metric performance scores}
\item{healthy_value}{A categorical value that denotes whether high or low values of a metric indicate a healthy stream. It can take the following values: \emph{high}, \emph{low}, or \emph{non-linear}}
\item{npsfm}{A Boolean variable denoting if a metric is part of the National Policy Statement for Freshwater Management 2020}
\item{suitability}{A numerical value (\emph{1}, \emph{2}, or \emph{3}) for each metric, assigned using expert assessment to quantify if metrics are fit for purpose. These values are used in the data integration process}
\item{key_metric}{A Boolean variable denoting if a metric is necessary for a holistic assessment of a river or stream. Metrics denoted as key metrics will be used to calculate plotting ratios of each indicator}
}
}
\usage{
reference_table_default
}
\description{
A dataset providing default values for excellent and poor ecological conditions for metrics used to calculate performance scores.
}
\keyword{datasets}
