% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riverhealth-package.R
\docType{package}
\name{riverhealth-package}
\alias{riverhealth}
\alias{riverhealth-package}
\title{riverhealth: New Zealand River Ecosystem Health Score}
\description{
Convinient functions to calculate a New Zealand River Ecosystem Health Score as proposed by Clapcott et al. (2019). The score serves as a simple and holistic measure of the biophysical condition of rivers and streams in New Zealand. It is based on the Freshwater Biophysical Ecosystem Health Framework and evaluates five core components of ecosystem health: Aquatic Life, Physical Habitat, Water Quality, Water Quantity, and Ecological Processes.
}
\author{
\strong{Maintainer}: Alba Cervantes-Loreto \email{alba.cervantesloreto@cawthron.org.nz} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
