% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_clean.R
\name{theme_clean}
\alias{theme_clean}
\title{Clean Theme for ggplot2 Plots}
\usage{
theme_clean()
}
\value{
A ggplot2 theme object with customized settings.
}
\description{
This function defines a custom ggplot2 theme that enhances the appearance of plots
by removing minor grid lines, adjusting strip text size and alignment, and setting
a background color for strip panels.
}
