% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_report_card.R
\name{plot_report_card}
\alias{plot_report_card}
\title{Plot report cards for multiple reporting scales}
\usage{
plot_report_card(
  component_scores,
  reference_table,
  npsfm_only = FALSE,
  font_size = 12,
  color_table = NULL,
  start_year = NULL,
  end_year = NULL
)
}
\arguments{
\item{component_scores}{A list containing data frames of scores for each component.}

\item{reference_table}{A reference table used for plotting.}

\item{npsfm_only}{Logical indicating whether to use only NPSFM components.}

\item{font_size}{Font size for the plots (default is 15).}

\item{color_table}{Optional color table for customizing plot colors.}

\item{start_year}{Optional start year for filtering data.}

\item{end_year}{Optional end year for filtering data.}
}
\value{
A list of plots, each corresponding to a different reporting scale.
}
\description{
This function generates plots for various environmental components based on provided scores and reference data.
}
\examples{
 macroinvertebrates <- data.frame( site = 1:3,class = c("universal"),
 indicator = "macroinvertebrates",
 component = "aquatic_life",
 reporting_scale = c("region_1","region_2","region_1"),
 mci = c(89,120,135))

 score_table <-calculate_health_score(indicators = list(macroinvertebrates),
 reference_table = reference_table_default)

 color_table <- c("Z" = "white",
 "W" = "#bdbcbc",
 "A" = "#20a7ad",
 "B+" = "#85bb5b",
 "B-" = "#85bb5b",
 "C+" = "#ffa827",
 "C-" = "#ffa827",
 "D" = "#e85129")

 plot_report_card(component_scores = list(score_table),
 reference_table = reference_table_default,
 color_table = color_table)
}
